requirements.txt
Flask==1.1.1
python-barcode==0.11.0
Pillow==6.2.1
scikit-image==0.17.2
gunicorn==20.0.4
gradio